export driver downloader logic and test to its own class
